add Solidity snippets, seed tokenization "as-is"
The majority of tokenization by the current mode-solidity.js is
correct, so we can use that to seed the files and data structures to
hold the tokenization data. After this step, only the current mistakes like
missing keyword tokenizations need to get corrected instead of creating
the whole tokenization corpus from scratch.

EVM assembly will be left out for now, in order to progress with the main project.
It can be seen as another language anyways. For situations with a
language-in-another-language (like PHP in HTML, or SQL in PHP), the ACE
project encourages to import a separate syntax highlighter module.